aLSO, IN THE 'pLOT OPTIONS BY CELL STATE' EXPANDER IN THErEPORTS TAB: tHE FIRST CELL POP SHOULD BE SELECTED BY DEFAULT.

git-tfs-id: [http://tfs.bumc.bu.edu:8080/tfs/BU-MICRO-MSI]$/Daphne/Daphne-skg;C6081

+++ Daphne/Simulation.cs
@@ -1170,7 +1170,7 @@ namespace Daphne
             }
             // 5., find the maximum force
             f_max = 0;
-            foreach (KeyValuePair<int, Pair> kvp in collisionManager.Pairs)
+            foreach (KeyValuePair<long, Pair> kvp in collisionManager.Pairs)
             {
                 if (kvp.Value.Force > f_max)
                 {
